docs(getting_started): add steps three and four for users to follow
Added steps for users to follow; adding the LiveStax CSS from the CDN,
and adding the UI elements for the application. Provided the user with
links to the LiveStax theme page, reference to MDN on HTML elements and
Bootstrap documentation on its grid system. Provided user a note on
using a placeholder image, informing them that `placehold.it` is only
one resource and they can use any other means to provide a placeholder
image.