docs(api_request): add step five and six for users to follow
Added step for users to follow, adding the jQuery Script from the CDN
and adding the targeting classes to UI Elements. How to capture the
input box value and make the request to the API service. Provided a note
that we using jQuery is not essential and there are other libraries
available.

Displayed method of manipulating the DOM using the data from the API
request with jQuery. Reiterated that they should have free choice of
technologies.